Add files via upload
This is MLP_conc.py, the main piece of code. This code computes basic data analyses, plots the correlation matrix, and then uses scikit learn's MLPClassifier to classify the strength categories of concrete mixes. After that is complete, the accuracy is reported, the loss curve is plotted, and the confusion matrix is plotted for the data. This piece of code references (1) the Excel spreadsheet, Concrete.xlsx and (2) a module created mainly from a course I took with Dr. Thomas Goebel at The University of Memphis, ml_utils.py (The only function used from this module is plot_confusion_matrix).